✨ feat: Add ObjectExtensions class with EqualsToAll and EqualsToAny methods
This commit introduces the `ObjectExtensions` class, which provides extension methods for the `object` class. The two new methods, `EqualsToAll` and `EqualsToAny`, allow developers to easily compare an object with an array of objects and determine if the object is equal to all or any of the objects in the array.