✨ feat: Add QuaternionExtensions class and meta file
The changes introduce a new `QuaternionExtensions` class that provides a set of extension methods for the `Quaternion` struct. These methods allow for easy manipulation of individual components of the quaternion, such as setting the values of specific axes or combinations of axes. The changes also include the addition of a meta file for the `QuaternionExtensions.cs` script.